Fix many compiler warnings
* 'load_succeeded', 'server' and 'priv' are all unused.
* The compiler believes 'cache_size' may be used uninitialized. In fact,
  that's not true, but the compiler doesn't know the relationship
  between a function returning FALSE (which skips one initialization
  path) implying error != NULL (which leads to the other initialization
  path).
* goffset is a signed 64-bit int, but circular-file stores the max file
  size as an unsigned 64-bit int.
* Two 'int i' iterator variables don't match the sign of the 'gsize'
  used in the termination condition.
* The compiler complains about old-style function definitions when a
  no-args definition doesn't have '(void)'.
* The compiler complained that emer_persistent_cache_new_full() might be
  a candidate for G_GNUC_NORETURN. Assuming assertions are enabled at
  compile-time, that's true, but if they were disabled it would return
  junk since there was no 'return' statement. Instead, explicitly return
  NULL, while logging at level CRITICAL (which is fatal in the test
  suite, which is all we really care about).
* tests: fix passing constant strings as non-const char *
* tests: fix GError.code sign warnings. In both cases the code is
  enum-typed, and so unsigned, but GError.code is a signed integer.

+++ daemon/emer-daemon.c
@@ -291,7 +291,7 @@ remove_events (EmerDaemon *self,
   if (num_events == 0)
     return;
 
-  for (gint i = 0; i < num_events; i++)
+  for (gsize i = 0; i < num_events; i++)
     {
       GVariant *curr_variant = g_ptr_array_index (priv->variant_array, i);
       priv->num_bytes_buffered -= emer_persistent_cache_cost (curr_variant);
